requirements.txt
streamlit>=1.32
numpy>=1.24
scipy>=1.10
matplotlib>=3.8
wfdb>=4.1
tensorflow>=2.14
xgboost>=1.7
scikit-learn>=1.3
joblib>=1.3
dill>=0.3.8
lightgbm>=4.3
catboost>=1.2
opencv-python>=4.9
PyWavelets>=1.6
neurokit2>=0.2.7
google-genai
protobuf==5.29.6